Página de perfil de usuario criada

+++ PAGES/atividades_ao_ar_livre/delete/delete_atividadePHP.php
@@ -11,20 +11,31 @@
     if (isset($_GET['codigo'])){
         $codigo = $_GET['codigo'];               
     
-        $atividade = "SELECT * FROM TABATV WHERE TABATV_Codigo = $codigo";
-        $queryAtividade = $mysqli->query($atividade) or die("Falha na execução do código sql" . $mysqli->error);
+        //Verifica se tem inscritos na atividade ao ar livre
+        $inscritos = "SELECT * FROM PARATV WHERE TABATV_Codigo = $codigo";
+        $queryInscritos = $mysqli->query($inscritos) or die("Falha na execução do código sql" . $mysqli->error);
 
-        if ($queryAtividade->num_rows == 1){
+        if($queryInscritos->num_rows == 0){
+            $atividade = "SELECT * FROM TABATV WHERE TABATV_Codigo = $codigo";
+            $queryAtividade = $mysqli->query($atividade) or die("Falha na execução do código sql" . $mysqli->error);
 
-            $deleteAtividade = "DELETE FROM TABATV WHERE TABATV_Codigo = $codigo";
-            $queryDeleteAtividade = $mysqli->query($deleteAtividade) or die("Falha na execução do código sql" . $mysqli->error);
-            header ("Location: ../../home/home.php");
+            if ($queryAtividade->num_rows == 1){                                
+                $deleteAtividade = "DELETE FROM TABATV WHERE TABATV_Codigo = $codigo";
+                $queryDeleteAtividade = $mysqli->query($deleteAtividade) or die("Falha na execução do código sql" . $mysqli->error);
+                header ("Location: ../../home/home.php");
 
-        }else{
-            header ("Location: ../../home/home.php");
-        }
-        
+            }else{
+                header ("Location: ../../home/home.php");
+            }
+        }else{?>
+            <!--Redireciona para a Página anterior-->
+            <script>
+                alert('Atividade ao ar livre não pode ser apagada pois possuí incritos!');
+                window.location.href = "../../home/home.php";  
+            </script>
+        <?php
+        }               
     }else{
         header ("Location: ../../home/home.php");
     }    
-?>
+?>
